BE-9-add_company_actions
Create a model for invitations from a company to user and a model for requests from user to a company.
Update existing ones to ensure all the fields can be accessed.

Add routers for the new models.

Add services for the models to allow owner sending an invitations to the user. Cancel sent invitations.

User to sent requests and cancel them.

Company can accept or decline a request.
User can accept or decline invitation.

In case of `accept` user will be assigned to the company.

Created according repositories for all specific invitations and requests db operations.